docs/pyproject.toml
[tool.poetry]
authors = ["23people <hola@23people.io>"]
description = "The 23people's Company Handbook"
maintainers = [
  "Manuel Reyes <manuel@23people.io>",
  "Marcelo Ampuero <marcelo@23people.io>",
]
name = "23people-handbook-docs"
package-mode = true
packages = []
readme = "README.md"
repository = "https://github.com/23people-io/23people-handbook-docs"
version = "1.1.1"

[tool.poetry.dependencies]
llama-index-core = "^0.12.5"
llama-index-embeddings-openai = "^0.3.1"
llama-index-llms-anthropic = "^0.5.0"
llama-index-llms-openai = "^0.3.10"
llama-index-readers-file = "^0.4.1"
loguru = "^0.7.3"
mkdocs = {extras = ["i18n"], version = "^1.6.1"}
mkdocs-exclude = "^1.0.2"
mkdocs-git-committers-plugin = "^0.2.3"
mkdocs-git-committers-plugin-2 = "^2.4.1"
mkdocs-git-revision-date-localized-plugin = "^1.2.7"
mkdocs-material = "^9.5.34"
mkdocs-minify-plugin = "^0.8.0"
mkdocs-redirects = "^1.2.1"
mkdocs-version-annotations = "^1.0.0"
mkdocstrings = "^0.25.2"
mkdocstrings-python = "^1.10.9"
pydantic = "^2.10.3"
python = "^3.11.5"
python-dotenv = "^1.0.1"
streamlit = "^1.41.1"

[tool.pyright]
defineConstant = {DEBUG = true}
exclude = [
  "**/.venv",
  "**/__pycache__",
]
include = ["src"]
reportMissingImports = "error"
reportMissingTypeStubs = false

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
